+++ src/app/logical/response/response.spec.ts
@@ -1,0 +1,14 @@
+import { Response } from "../../logical/response/response";
+import { ResponseService } from "./response";
+
+describe("Response service", () => {
+
+    let service;
+
+    beforeEach(() => {
+        service = new ResponseService();
+    });
+
+    
+
+});
