+++ src/app/logical/response/response.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@angular/core';
+import { Response } from 'src/app/logical/response/response';
+
+@Injectable()
+export class ResponseService {
+
+}
