+++ src/app/services/services.module.ts
@@ -1,3 +1,2 @@
 export * from './api/api.service';
 export * from './page/page.service';
-export * from './response/response.service';
