--- src/app/util/response/response.ts
@@ -1,6 +1,0 @@
-import { Observable } from 'rxjs';
-
-export class Response {
-
-    constructor() {}
-}
